Home db cache (#833)
* Refactor auth tokens

* Initial implementation of cache

* Add supported feature for Bolt 5.8 and fix cache key for home database

* Enhance routing and connection pool interfaces to support home database caching and SSR configuration

* Fix for Stefan's category/classification issue in WithNotifications, some tidying, moved GQL Notifications out of preview

* Add test for WithNotifications to set multiple categories and classifications

* Refactor NotificationsConfig to handle null disabled categories and classifications

* Update tests to assert DisabledCategories is empty instead of null

* Cache key now working?

* Refactor authentication token handling and update tests for response handlers

* Get unit tests working

* pre cache refactor

* Refactor logging output format and introduce ConsoleTextWriter for enhanced console output

* testkit tests looking good

* Enhance BeginResponseHandler and RunResponseHandler to support default database flag

* Unit tests all fixed

* Review notes bugs fixed

* HomeDbCache cache purging

* Add support for Bolt protocol version 5.8

* fix failing tests

* maybe fix testkit?

* Add exception detail to aid in debugging testkit failures

* Fix failing testkit tests

* Implement IsDirectDriver property in connection provider and related classes

* LoadBalancer fix for mixed clusters

* Fix logic in "dance"

* fix testkit tests?

+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/LambdaMappingRecordExtensionsTests.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) "Neo4j"
 // Neo4j Sweden AB [https://neo4j.com]
-//
+// 
 // Licensed under the Apache License, Version 2.0 (the "License").
 // You may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
+// 
 //     http://www.apache.org/licenses/LICENSE-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -284,7 +284,7 @@ public class LambdaMappingRecordExtensionsTests
         result.e.Should().Be(123L);
         result.f.Should().Be('x');
         result.g.Should().Be(123.45m);
-        result.h.Should().Be((byte)123);
+        result.h.Should().Be(123);
     }
 
     [Fact]
@@ -349,8 +349,8 @@ public class LambdaMappingRecordExtensionsTests
         result.e.Should().Be(123L);
         result.f.Should().Be('x');
         result.g.Should().Be(123.45m);
-        result.h.Should().Be((byte)123);
-        result.i.Should().Be((short)12345);
+        result.h.Should().Be(123);
+        result.i.Should().Be(12345);
     }
 
     [Fact]
@@ -420,9 +420,9 @@ public class LambdaMappingRecordExtensionsTests
         result.e.Should().Be(123L);
         result.f.Should().Be('x');
         result.g.Should().Be(123.45m);
-        result.h.Should().Be((byte)123);
-        result.i.Should().Be((short)12345);
-        result.j.Should().Be((ushort)12345);
+        result.h.Should().Be(123);
+        result.i.Should().Be(12345);
+        result.j.Should().Be(12345);
     }
 
     [Fact]
